Move deploy candidates to deploy handler and share between autopromoter and autorollback
