+++ deploy-service/common/src/main/java/com/pinterest/deployservice/handler/DeployHandler.java
@@ -18,6 +18,7 @@ package com.pinterest.deployservice.handler;
 import com.pinterest.deployservice.ServiceContext;
 import com.pinterest.deployservice.bean.AcceptanceStatus;
 import com.pinterest.deployservice.bean.BuildBean;
+import com.pinterest.deployservice.bean.BuildTagBean;
 import com.pinterest.deployservice.bean.CommitBean;
 import com.pinterest.deployservice.bean.DeployBean;
 import com.pinterest.deployservice.bean.DeployFilterBean;
@@ -31,7 +32,10 @@ import com.pinterest.deployservice.bean.PromoteDisablePolicy;
 import com.pinterest.deployservice.bean.PromoteType;
 import com.pinterest.deployservice.bean.ScheduleBean;
 import com.pinterest.deployservice.bean.ScheduleState;
+import com.pinterest.deployservice.bean.TagValue;
 import com.pinterest.deployservice.bean.UpdateStatement;
+import com.pinterest.deployservice.buildtags.BuildTagsManager;
+import com.pinterest.deployservice.buildtags.BuildTagsManagerImpl;
 import com.pinterest.deployservice.common.CommonUtils;
 import com.pinterest.deployservice.common.Constants;
 import com.pinterest.deployservice.common.DeployInternalException;
@@ -52,11 +56,15 @@ import com.google.common.base.Joiner;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.dbcp.BasicDataSource;
 import org.apache.commons.lang.StringUtils;
+import org.joda.time.Interval;
+import org.joda.time.format.ISODateTimeFormat;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -84,6 +92,7 @@ public class DeployHandler implements DeployHandlerInterface{
     private ExecutorService jobPool;
     private String deployBoardUrlPrefix;
     private String changeFeedUrl;
+    private BuildTagsManager buildTagsManager;
 
     private final class NotifyJob implements Callable<Void> {
         private EnvironBean envBean;
@@ -180,6 +189,7 @@ public class DeployHandler implements DeployHandlerInterface{
         jobPool = serviceContext.getJobPool();
         deployBoardUrlPrefix = serviceContext.getDeployBoardUrlPrefix();
         changeFeedUrl = serviceContext.getChangeFeedUrl();
+        buildTagsManager = new BuildTagsManagerImpl(serviceContext.getTagDAO());
     }
 
     private String generateMentions(EnvironBean envBean, DeployBean newDeployBean, DeployBean oldDeployBean) {
@@ -449,5 +459,52 @@ public class DeployHandler implements DeployHandlerInterface{
             updateScheduleBean.setState_start_time(System.currentTimeMillis());
             scheduleDAO.update(updateScheduleBean, scheduleId);   
         }
-    } 
+    }
+
+    public List<DeployBean> getDeployCandidates(String envId, Interval interval, int size, boolean onlyGoodBuilds) throws Exception {
+        LOG.info("Search Deploy candidates between {} and {} for environment {}",
+            interval.getStart().toString(ISODateTimeFormat.dateTime()),
+            interval.getEnd().toString(ISODateTimeFormat.dateTime()),
+            envId);
+        List<DeployBean> taggedGoodDeploys = new ArrayList<DeployBean>();
+
+        List<DeployBean> availableDeploys = deployDAO.getAcceptedDeploys(envId, interval, size);
+
+        if (!onlyGoodBuilds) {
+            return availableDeploys;
+        }
+
+        if(!availableDeploys.isEmpty()) {
+            Map<String, DeployBean> buildId2DeployBean = new HashMap<String, DeployBean>();
+            for(DeployBean deployBean: availableDeploys) {
+                String buildId = deployBean.getBuild_id();
+                if(StringUtils.isNotEmpty(buildId)) {
+                    buildId2DeployBean.put(buildId, deployBean);
+                }
+            }
+            List<BuildBean> availableBuilds = buildDAO.getBuildsFromIds(buildId2DeployBean.keySet());
+            List<BuildTagBean> buildTagBeanList = buildTagsManager.getEffectiveTagsWithBuilds(availableBuilds);
+            for(BuildTagBean buildTagBean: buildTagBeanList) {
+                if(buildTagBean.getTag() != null && buildTagBean.getTag().getValue() == TagValue.BAD_BUILD) {
+                    // bad build,  do not include
+                    LOG.info("Env {} Build {} is tagged as BAD_BUILD, ignore", envId, buildTagBean.getBuild());
+                } else {
+                    String buildId = buildTagBean.getBuild().getBuild_id();
+                    taggedGoodDeploys.add(buildId2DeployBean.get(buildId));
+                }
+            }
+        }
+        // should order deploy bean by start date desc
+        if(taggedGoodDeploys.size() > 0) {
+            Collections.sort(taggedGoodDeploys, new Comparator<DeployBean>() {
+                @Override
+                public int compare(final DeployBean d1, final DeployBean d2) {
+                    return Long.compare(d2.getStart_date(), d1.getStart_date());
+                }
+            });
+            LOG.info("Env {} the first deploy candidate is {}", envId, taggedGoodDeploys.get(0).getBuild_id());
+        }
+        return taggedGoodDeploys;
+    }
+
 }
